Fix broken varints test + use int32 instead of uint32 for varints + updated some varint-related function param types to be more consistent

+++ protocol/packet/play/chunk_data.go
@@ -18,6 +18,6 @@ func (c *OutChunkData) MarshalPacket(w *protocol.WriteBuffer) {
 	w.WriteUint32(uint32(c.Z))
 	w.WriteBool(c.FullChunk)
 	w.WriteUint16(c.Mask)
-	w.WriteVarInt(len(c.Data))
+	w.WriteVarInt(int32(len(c.Data)))
 	w.WriteBytes(c.Data)
 }
